move aggregate_error pkg to internal
